# Cache responses in ApiSwaggerTest