### What is this PR trying to achieve?

Make our tests go faster, and make adding extra Swagger tests cheaper.

In a local IntelliJ run, this got the overall test suite from ~27s to ~16s (saving 11s). More significantly, the cost of everything but the first test goes from ~1.5s to ~1ms, so it’s much cheaper to add new tests.

### Who is this change for?

㊙️ 🔒